refactored favorite component; added to individual pages
